feat: UserService 추상화
